backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0 
django-cors-headers==4.3.1
Pillow==10.1.0
